working refactor to vector search app
