+++ learning_resources_search/serializers.py
@@ -469,16 +469,6 @@ class LearningResourcesSearchRequestSerializer(SearchRequestSerializer):
     )
 
 
-class LearningResourcesVectorSearchRequestSerializer(serializers.Serializer):
-    q = serializers.CharField(required=False, help_text="The search text")
-    offset = serializers.IntegerField(
-        required=False, help_text="The initial index from which to return the results"
-    )
-    limit = serializers.IntegerField(
-        required=False, help_text="Number of results to return per page"
-    )
-
-
 class ContentFileSearchRequestSerializer(SearchRequestSerializer):
     id = serializers.ListField(
         required=False,
@@ -649,21 +639,6 @@ class PercolateQuerySerializer(serializers.ModelSerializer):
     class Meta:
         model = PercolateQuery
         exclude = (*COMMON_IGNORED_FIELDS, "users")
-
-
-class LearningResourcesVectorSearchResponseSerializer(SearchResponseSerializer):
-    @extend_schema_field(LearningResourceSerializer(many=True))
-    def get_results(self, instance):
-        return instance.get("hits", {})
-
-    def get_count(self, instance) -> int:
-        return instance.get("total", {}).get("value")
-
-    def get_metadata(self, _) -> SearchResponseMetadata:
-        return {
-            "aggregations": [],
-            "suggest": [],
-        }
 
 
 class LearningResourcesSearchResponseSerializer(SearchResponseSerializer):
