+++ learning_resources_search/views.py
@@ -21,7 +21,6 @@ from learning_resources_search.api import (
     execute_learn_search,
     subscribe_user_to_search_query,
     unsubscribe_user_from_percolate_query,
-    vector_search,
 )
 from learning_resources_search.constants import CONTENT_FILE_TYPE, LEARNING_RESOURCE
 from learning_resources_search.models import PercolateQuery
@@ -30,8 +29,6 @@ from learning_resources_search.serializers import (
     ContentFileSearchResponseSerializer,
     LearningResourcesSearchRequestSerializer,
     LearningResourcesSearchResponseSerializer,
-    LearningResourcesVectorSearchRequestSerializer,
-    LearningResourcesVectorSearchResponseSerializer,
     PercolateQuerySerializer,
     PercolateQuerySubscriptionRequestSerializer,
 )
@@ -86,53 +83,6 @@ class LearningResourcesSearchView(ESView):
                 return Response(response)
             else:
                 response = LearningResourcesSearchResponseSerializer(
-                    response, context={"request": request}
-                ).data
-                response["results"] = list(response["results"])
-                return Response(response)
-        else:
-            errors = {}
-            for key, errors_obj in request_data.errors.items():
-                if isinstance(errors_obj, list):
-                    errors[key] = errors_obj
-                else:
-                    errors[key] = list(set(chain(*errors_obj.values())))
-            return Response(errors, status=400)
-
-
-@method_decorator(blocked_ip_exempt, name="dispatch")
-@extend_schema_view(
-    get=extend_schema(
-        parameters=[LearningResourcesVectorSearchRequestSerializer()],
-        responses=LearningResourcesVectorSearchResponseSerializer(),
-    ),
-)
-@action(methods=["GET"], detail=False, name="Search Learning Resources")
-class LearningResourcesVectorSearchView(ESView):
-    """
-    Vector Search for learning resources
-    """
-
-    permission_classes = ()
-
-    @method_decorator(
-        cache_page_for_anonymous_users(
-            settings.SEARCH_PAGE_CACHE_DURATION, cache="redis", key_prefix="search"
-        )
-    )
-    @extend_schema(summary="Vector Search")
-    def get(self, request):
-        request_data = LearningResourcesVectorSearchRequestSerializer(data=request.GET)
-
-        if request_data.is_valid():
-            query_text = request_data.data.get("q", "")
-            limit = request_data.data.get("limit", 10)
-            offset = request_data.data.get("offset", 0)
-            response = vector_search(query_text, limit=limit, offset=offset)
-            if request_data.data.get("dev_mode"):
-                return Response(response)
-            else:
-                response = LearningResourcesVectorSearchResponseSerializer(
                     response, context={"request": request}
                 ).data
                 response["results"] = list(response["results"])
