+++ main/urls.py
@@ -46,6 +46,7 @@ urlpatterns = (
         re_path(r"", include("profiles.urls")),
         re_path(r"", include("embedly.urls")),
         re_path(r"", include("learning_resources_search.urls")),
+        re_path(r"", include("vector_search.urls")),
         re_path(r"", include("ckeditor.urls")),
         re_path(r"", include("widgets.urls")),
         re_path(r"", include("openapi.urls")),
