+++ vector_search/models.py
@@ -1,0 +1,1 @@
+# Create your models here.
